+++ src/rigyard/providers/docker/__init__.py
@@ -4,11 +4,13 @@ from .build_backend import DockerExecBuildBackend
 from .container_backend import DockerContainerBackend
 from .containers import docker_container_sources, docker_containers_source
 from .image_backend import DockerImageBackend
+from .test_backend import DockerExecTestBackend
 
 __all__ = [
     "DockerContainerBackend",
     "DockerExecBuildBackend",
     "DockerImageBackend",
+    "DockerExecTestBackend",
     "docker_container_sources",
     "docker_containers_source",
 ]